requirements.txt
fastapi
uvicorn[standard]
gunicorn
apscheduler
httpx
curl_cffi
rapidfuzz
python-dotenv
unidecode
requests
postgrest
PyJWT>=2.8.0
